+++ src/Servant/Hateoas/ContentType/HAL.hs
@@ -1,9 +1,11 @@
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE UndecidableInstances #-}
-{-# LANGUAGE DefaultSignatures #-}
 
 module Servant.Hateoas.ContentType.HAL
-( HAL
+(
+  -- * Content-Type
+  HAL
+
+  -- * Resource-Type
 , HALResource(..)
 )
 where
@@ -11,62 +13,52 @@ where
 import Servant.Hateoas.Resource
 import Servant.API.ContentTypes
 import qualified Network.HTTP.Media as M
-import Servant.Links
 import qualified Data.Foldable as Foldable
 import Data.Some.Constraint
 import Data.Kind
-import Data.Proxy
 import Data.Aeson
 import Data.Aeson.KeyMap (singleton)
 import GHC.Exts
-import GHC.TypeLits
 import GHC.Generics
-import GHC.Records
 
--- | Data-Kind representing Content-Types of Hypertext Application Language (HAL).
+-- | Type representing Content-Types of Hypertext Application Language (HAL).
 --
 --   Type parameter @t@ is the mime type suffix in @application/hal+t@.
 data HAL (t :: Type)
 
--- | Resource wrapper for HAL.
+type instance MkResource (HAL t) = HALResource
+
+-- | HAL-resource representation.
 data HALResource a = HALResource
   { resource :: a                                       -- ^ Wrapped resource
-  , links    :: [(String, Link)]                        -- ^ Pairs @(rel, link)@ for relations
+  , rels     :: [(String, ResourceLink)]                -- ^ Pairs @(rel, link)@ for hypermedia relations
   , embedded :: [(String, SomeF HALResource ToJSON)]    -- ^ Pairs @(rel, resource)@ for embedded resources
-  } deriving (Generic)
+  } deriving (Generic, Functor)
 
 instance Resource HALResource where
-  addLink l (HALResource r ls es) = HALResource r (l:ls) es
+  wrap x = HALResource x [] []
+  addRel l (HALResource r ls es) = HALResource r (l:ls) es
 
 instance Accept (HAL JSON) where
   contentType _ = "application" M.// "hal+json"
 
-instance ToJSON a => MimeRender (HAL JSON) (HALResource a) where
+instance ToJSON (HALResource a) => MimeRender (HAL JSON) (HALResource a) where
   mimeRender _ = encode
 
 instance {-# OVERLAPPABLE #-} ToJSON a => ToJSON (HALResource a) where
-  toJSON (HALResource res ls es) = case toJSON res of
-    Object kvm -> Object $ (singleton "_links" ls') <> (singleton "_embedded" es') <> kvm
-    v -> v
+  toJSON (HALResource res ls es) = Object $ (singleton "_links" ls') <> (singleton "_embedded" es') <> (case toJSON res of Object kvm -> kvm ; _ -> mempty)
     where
-      ls' = object [fromString rel .= object ["href" .= linkURI href] | (rel, href) <- ls]
+      ls' = object [fromString rel .= object ["href" .= href] | (rel, href) <- ls]
       es' = object [fromString name .= toJSON e | (name, (Some1 e)) <- es]
 
-instance {-# OVERLAPPING #-} (ToJSON a, Related a, KnownSymbol (CollectionName a)) => ToJSON [HALResource a] where
-  toJSON xs = object ["_links" .= (mempty :: Object), "_embedded" .= es]
+instance {-# OVERLAPPING #-} ToJSON a => ToJSON (HALResource [a]) where
+  toJSON (HALResource xs ls es) = object ["_links" .= ls', "_embedded" .= object (exs <> es')]
     where
-      es = object $
-        [  fromString (symbolVal (Proxy @(CollectionName a)))
-        .= (Array $ Foldable.foldl' (\xs' x -> xs' <> pure (toJSON x)) mempty xs)
-        ]
+      ls' = object [fromString rel .= object ["href" .= href] | (rel, href) <- ls]
+      es' = fmap (\(eName, (Some1 e)) -> fromString eName .= toJSON e) es
+      exs = [ "items"
+              .= (Array $ Foldable.foldl' (\xs' x -> xs' <> pure (toJSON x)) mempty xs)
+            ]
 
 instance EmbeddingResource HALResource where
-  embed e (HALResource r ls es) = HALResource r ls $ fmap (\res -> Some1 $ HALResource res [] []) e : es
-
-instance {-# OVERLAPPABLE #-}
-  ( Related a, HasField (IdSelName a) a id, IsElem (GetOneApi a) api
-  , HasLink (GetOneApi a), MkLink (GetOneApi a) Link ~ (id -> Link)
-  , BuildRels api (Relations a) a
-  , Resource HALResource
-  ) => ToResource api HALResource a where
-  toResource x = HALResource x (defaultLinks (Proxy @api) x) mempty
+  embed e (HALResource r ls es) = HALResource r ls $ fmap Some1 e : es