+++ CHANGELOG.md
@@ -6,6 +6,17 @@ file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PVP versioning](https://pvp.haskell.org/).
 
+## v0.3.0 _(2024-12-24)_
+
+### Added
+- Added functionality to derive a HATEOAS Layer-Api and rewritten HATEOAS API from your API and their respective server-implementations.
+
+### Changed
+- Removed argument `api` from class `ToResource`
+
+### Removed
+- Temporarily removed support for content-type `application/vnd.collection+json`
+
 ## v0.2.2 _(2024-12-01)_
 
 ### Changed